102-fibonacci.c

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,8 +8,8 @@
 
 int main(void)
 {
-	int term1 = 1;
-	int term2 = 2;
+	unsigned long int term1 = 1;
+	unsigned long int term2 = 2;
 	int next;
 	int i;
